% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_print.R
\name{df_print}
\alias{df_print}
\title{Fonction pour définir le dataframe qui s'affiche pour l'utilisateur}
\usage{
df_print(df = app.voitures::voitures_2014)
}
\arguments{
\item{df}{}
}
\value{
un dataframe avec des jolis noms de variable
}
\description{
Fonction pour définir le dataframe qui s'affiche pour l'utilisateur
}
