% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_conso.R
\name{predict_conso}
\alias{predict_conso}
\title{Fonction entrapinant le modèle permettant de prédire le niveau de consommation}
\usage{
predict_conso(train = app.voitures::df_print())
}
\arguments{
\item{train}{table sur laquelle on entraîne le modèle}
}
\value{
un modèle qu'on peut utiliser pour la prédiction
}
\description{
Fonction entrapinant le modèle permettant de prédire le niveau de consommation
}
