% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voitures_2014.R
\docType{data}
\name{voitures_2014}
\alias{voitures_2014}
\title{Caractéristiques de voitures en France}
\format{A data frame with 55044 rows and 26 variables:}
\source{
\url{https://www.data.gouv.fr/fr/datasets/emissions-de-co2-et-de-polluants-des-vehicules-commercialises-en-france/}
}
\usage{
voitures_2014
}
\description{
Un dataframe avec les caractéristiques des voitures en France
et leurs consommations. Données 20014
}
\keyword{datasets}
